refactor: consolidate error handling and validation in CustomFieldsController
- Removed the customFieldsHelpers module and integrated its functionality directly into the CustomFieldsController.
- Updated methods to use class methods for validation and error handling, improving code organization and readability.
- Enhanced error management for database and PagerDuty interactions, providing clearer responses for various error scenarios.
- Added private methods for input validation, sanitization, and description normalization to streamline custom field processing.
